feat: complete chula sso
