+++ lighthouserc.js
@@ -6,7 +6,7 @@ module.exports = {
                 chromeFlags: "--ignore-certificate-errors"
             },
             numberOfRuns: 3,
-            url: ["https://127.0.0.1:8000/", "https://127.0.0.1:8000/test/", "https://127.0.0.1:8000/event/", "https://127.0.0.1:8000/user-checkin/"]
+            url: ["https://127.0.0.1:8000/", "https://127.0.0.1:8000/test/", "https://127.0.0.1:8000/event/", "https://127.0.0.1:8000/user-checkin/", "https://127.0.0.1:8000/login/"]
         },
         upload: {
             target: 'lhci',
