+++ src/apollo/client.ts
@@ -1,11 +1,13 @@
 import { ApolloClient } from "apollo-client";
 import { createHttpLink } from "apollo-link-http";
 import { InMemoryCache } from "apollo-cache-inmemory";
+import { AUTH_KEY } from "@/modules/home/components/signup/constant";
 
 // HTTP connection to the API
 const httpLink = createHttpLink({
   // You should use an absolute URL here
-  uri: "https://api.onepass.app/graphql"
+  uri: "https://api.onepass.app/graphql",
+  headers: { authorization: `Bearer ${window.localStorage.getItem(AUTH_KEY)}` }
 });
 
 // Cache implementation
