+++ src/modules/authentication/constant.ts
@@ -1,0 +1,1 @@
+export const userSymbol = Symbol("user");
