+++ src/modules/home/components/signup/constant.ts
@@ -1,0 +1,1 @@
+export const AUTH_KEY = "token";
