+++ src/router/index.ts
@@ -1,33 +1,45 @@
 import { createRouter, createWebHistory } from "vue-router";
 
-import Home from "@/modules/home/Home.vue";
-const Test = () => import("@/modules/authentication/TestPage.vue");
+const Home = () => import("@/modules/home/Home.vue");
+const Test = () => import("@/modules/test/TestPage.vue");
 const EventInfo = () => import("@/modules/eventInfo/EventInfoPage.vue");
 const Wallet = () => import("@/modules/wallet/WalletPage.vue");
 const UserCheckIn = () => import("@/modules/userCheckIn/UserCheckInPage.vue");
+const Login = () => import("@/modules/login/Login.vue");
+const PageSkeleton = () => import("@/commons/components/PageSkeleton.vue");
 
 const router = createRouter({
   history: createWebHistory(),
   routes: [
     {
       path: "/",
-      component: Home
+      component: PageSkeleton,
+      children: [
+        {
+          path: "",
+          component: Home
+        },
+        {
+          path: "/test",
+          component: Test
+        },
+        {
+          path: "/event",
+          component: EventInfo
+        },
+        {
+          path: "/wallet",
+          component: Wallet
+        },
+        {
+          path: "/user-checkin",
+          component: UserCheckIn
+        }
+      ]
     },
     {
-      path: "/test",
-      component: Test
-    },
-    {
-      path: "/event",
-      component: EventInfo
-    },
-    {
-      path: "/wallet",
-      component: Wallet
-    },
-    {
-      path: "/user-checkin",
-      component: UserCheckIn
+      path: "/login",
+      component: Login
     }
   ]
 });
